Readme, delete desktop, refactor
