+++ core/src/no/ntnu/game/Sound/SoundPlayer.java
@@ -1,4 +1,4 @@
-package no.ntnu.game.Settings;
+package no.ntnu.game.Sound;
 
 public interface SoundPlayer {
