--- desktop/src/no/ntnu/game/DesktopLauncher.java
@@ -1,15 +1,0 @@
-package no.ntnu.game;
-
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-
-
-// Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
-public class DesktopLauncher {
-	public static void main (String[] arg) {
-		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
-		config.setForegroundFPS(60);
-		config.setTitle("Star Knight");
-		new Lwjgl3Application(new StarKnight(new DesktopFirebase()), config);
-	}
-}
